### 2017-09-06 #395 @CBenghi
Hello @arfon,
thanks for your comment, I had concerns about this aspect as well, but reading again in your guidelines the parts that apply are highlighted as follows:

"JOSS publishes articles about research software. This definition includes software that: solves complex modeling problems in a scientific context (physics, mathematics, biology, medicine, social science, neuroscience, **engineering**); supports the functioning of research instruments or the execution of research experiments; **extracts knowledge from large data sets**; offers a mathematical library, or similar."

The libriary offers the ability to extract information from construction projects, and it's been used in the past to enable research projects on carbon assessment and several other aspects of construction management and construction engineering. It has also served in the areas of information science (I know at least of three publications in [Advanced Engineering Informatics](https://www.journals.elsevier.com/advanced-engineering-informatics) that have used it for their work, but in absence of a formal publication the usage has (sadly) not been acknowledged.

Unfortunately construction engineering publications have sometimes lower standards than other branches on this front, but my hope was that through Joss some of this might improve.

I have provided in the paper a list of government sponsored research projects, that this library supports, how can I complement this info to match the requirements?

Thanks,
Claudio
