### 2017-09-07 #395 @arfon
> I've read the requirements again and spotted omissions, I'm going to amend the paper and inform you when an update is available for consideration.

Excellent, many thanks!
