### 2017-09-08 1f4e47a
Updated JOSS paper version.
