+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Xbim.Essentials: a library for interoperable building information applications'
+title: 'Xbim.Essentials: full control over Building Information Models'
 tags:
   - BIM
   - BuildingSmart
@@ -23,35 +23,42 @@ bibliography: paper.bib
 
 # Summary
 
-The Xbim.Essentials library offers the open-source infrastructure to develop Building Information Modelling (BIM) solutions for commercial, research and open-source environments under the terms of the CDDL [@sun_microsystems_common_2005].
+The Xbim.Essentials library offers data extraction, data transformation and data validation functions for Building Information Models (BIM); its robust and optimised implementation allows efficient operations on gigabyte-range files for researchers and practitioners interested in the built environment.
 
-The library implements the complete object model of BuildingSMART's IFC schemas [@buildingsmart_international_summary_2017] along with APIs for their management in a production environment under the Microsoft .Net framework.
+The library implements the complete object model of BuildingSMART's IFC schemas [-@buildingsmart_international_summary_2017] along with APIs for their management under the terms of the CDDL [@sun_microsystems_common_2005], which makes it also suitable for commercial research and development projects.
 
-APIs are available for:
+Relevant fields of research span from social sciences to construction engineering and applied mathematics on account of the breadth of domains covered in the over 800 classes of the schema across building lifecycles.
 
-- Single point management of any IFC file, including XML, STEP21 and IFCZIP formats
-- Disk-cached and in-memory options to suit a range of deployment infrastructures
-- Federation, merging and deleting entities and sub-sets of models (a non-trivial function because of potential cyclic and bi-directional relationships)
-- Embedded Schema Metadata information for programmatic data extraction and presentation of IFC properties
-- Verification of model consistency through EXPRESS WHERE clauses and metadata constraints
-- Read and write c# interfaces across schema versions, allowing to develop one code base for multiple formats
-- log management for notification events where exceptions are not appropriate
+The project implements public APIs for:
 
-Additionally, the Xbim.IO.TableStore namespace supports similar APIs, where appropriate, on BuildingSMART's COBie object model [@buildingsmart_cobie_2013].
+- Federating, merging and splitting models and entities within models  (non-trivial because of cyclic and bi-directional relationships defined in the schemas).
+- Verification of data quality through EXPRESS WHERE clauses and metadata constraints.
+- programmatic management of IFC properties and relations through Schema Metadata.
+- Single point management of any IFC file, including XML, STEP21 and IFCZIP formats.
 
-Where possible, classes in the codebase are generated programmatically from the EXPRESS specification files [@iso_iso-10303-11:2004_2013] ensuring full compliance with the standards. Schemas currently implemented are:
+Architectural features of the solution include:
+
+- Full access to the models through C# interfaces, allowing schema-agnostic data logic on any supported version.
+- Disk-based and in-memory options to suit diverse of workflow infrastructures.
+- Log management for notification events where exceptions are not appropriate.
+
+Additionally, the Xbim.IO.TableStore namespace supports similar APIs, when conceptually suitable, on BuildingSMART's COBie object model [@buildingsmart_cobie_2013], extending the domain of relevance to construction operations and facility management.
+
+![Figure 1: Internal dependencies graph.](CodeDependenciesGraph.png)
+
[…]
