+++ paper/paper.bib
@@ -11,6 +11,28 @@
   groups = {Institutions and standards,Institutions and standards}
 }
 
+@misc{sun_microsystems_common_2005,
+  title = {Common {{Development}} and {{Distribution License}} 1.0},
+  url = {https://opensource.org/licenses/CDDL-1.0},
+  shorttitle = {{{CDDL}}-1.0},
+  timestamp = {2017-09-08T15:22:05Z},
+  publisher = {{Open Source Initiative}},
+  author = {{Sun Microsystems}},
+  urldate = {2017-09-04},
+  date = {2005-01},
+  groups = {Xbim,Xbim}
+}
+
+@online{buildingsmart_cobie_2013,
+  title = {{{COBie}} - {{Model}} View Definition},
+  url = {http://docs.buildingsmartalliance.org/MVD_COBIE/},
+  timestamp = {2017-09-04T17:54:02Z},
+  author = {{buildingSMART}},
+  urldate = {2017-08-07},
+  date = {2013-10-30},
+  groups = {COBie,Institutions and standards,COBie,Institutions and standards}
+}
+
 @online{model_support_group_ifc2x3_2007,
   title = {{{IFC2x3 Technical Corrigendum}} 1 - {{Final Documentation}}},
   url = {http://www.buildingsmart-tech.org/ifc/IFC2x3/TC1/html/index.htm},
@@ -18,33 +40,6 @@
   author = {{Model Support Group}},
   urldate = {2017-09-04},
   date = {2007},
-  groups = {Institutions and standards,Institutions and standards}
-}
-
-@misc{iso_iso-10303-11:2004_2013,
-  title = {{{ISO}}-10303-11:2004 {{Industrial}} Automation Systems and Integration -- {{Product}} Data Representation and Exchange -- {{Part}} 11: {{Description}} Methods: {{The EXPRESS}} Language Reference Manual},
-  url = {https://www.iso.org/standard/38047.html},
-  shorttitle = {{{EXPRESS}} Data Modeling Language},
-  abstract = {ISO 10303 specifies a language by which aspects of product data can be defined. The language is called EXPRESS. ISO 10303-11:2004 also specifies a graphical representation for a subset of the constructs in the EXPRESS language. This graphical representation is called EXPRESS-G. EXPRESS is a data specification language as defined in ISO 10303-1. It consists of language elements that allow an unambiguous data definition and specification of constraints on the data defined.
-
-The following are within the scope of ISO 10303-11:2004:
-
-data types;
-constraints on instances of the data types.
-The following are outside the scope of this part of ISO 10303:
-
-definition of database formats;
-definition of file formats;
-definition of transfer formats;
-process control;
-information processing;
-exception handling.
-EXPRESS is not a programming language.},
-  timestamp = {2017-09-04T16:23:07Z},
-  publisher = {{International Organization for Standardization}},
-  author = {{ISO}},
-  urldate = {2017-09-04},
-  date = {2013-04},
   groups = {Institutions and standards,Institutions and standards}
 }
 
@@ -58,24 +53,13 @@ EXPRESS is not a programming language.},
   groups = {Institutions and standards,Institutions and standards}
 }
 
-@online{buildingsmart_cobie_2013,
-  title = {{{COBie}} - {{Model}} View Definition},
-  url = {http://docs.buildingsmartalliance.org/MVD_COBIE/},
-  timestamp = {2017-09-04T17:54:02Z},
-  author = {{buildingSMART}},
-  urldate = {2017-08-07},
[…]
