### 2017-09-08 #395 @CBenghi
Hello @arfon,
I've revised the structure of the paper to explicit the areas of research that could benefit from the software and clarify the features.
I've also re-framed the rest of the paper to prioritise the abstract capabilities rather than API implementation.

I hope you'll find the changes sufficient to progress to review.

Best,
Claudio
