### 2017-09-11 #395 @arfon
> I hope you'll find the changes convincing to progress to review.

Many thanks for these updates @CBenghi. Before we can move to review we'll need to have one of the @openjournals/joss-editors to volunteer to edit this submission.

In the meantime, if you have any suggestions for potential reviewers please mention them here.
