### 2017-09-12 #395 @CBenghi
Thanks for your note, @arfon.

Having had a look at the list of editors I am uncertain about ideal reviewer.

I haven't spotted anybody with a connected research background or anyone with a c# or .net repository on their github page.

If you can find any volunteer I'd be happy to guide them meticulously through any technical hurdles.

I'm confident that the project is set up to production standards. As long as they have access to Visual studio, building and testing the solution, as well as installing the binaries as a dependency for setting up custom analysis scenarios, should take only a few moments.

For any future needs, of course, I'd be happy to volunteer for the review of engineering related submissions or .Net implementations.
