### 2017-09-12 8aca336
Attempt to resolve net47 zip file bindings
