### 2017-09-13 db20f1e
New method of EntityCollection fixed to handle null initialisation function
