### 2017-09-13 20307b0
WIP: Memory model reading partial data.
