### 2017-09-13 c98072f
WIP: Memory model reading partial data.
