### 2017-09-13 9ecd083
MemoryModel has a new function LoadStep21Part which takes partial file and loads it into current model. This is usefull for partial data exchange.
