+++ Xbim.IO.MemoryModel/MemoryModel/MemoryModel.cs
@@ -12,6 +12,7 @@ using Xbim.Common.Geometry;
 using Xbim.Common.Metadata;
 using Xbim.Common.Step21;
 using Xbim.IO.MemoryModel;
+using Xbim.IO.Parser;
 using Xbim.IO.Step21;
 using Xbim.IO.Xml;
 using Xbim.IO.Xml.BsConf;
@@ -27,8 +28,8 @@ namespace Xbim.IO.Memory
         {
             using (var zipStream = new ZipArchive(fileStream))
             {
-                return zipStream.Entries.FirstOrDefault(z=>z.Name.IsStepTextFile()); //ignores xbim and zip files
-            }          
+                return zipStream.Entries.FirstOrDefault(z => z.Name.IsStepTextFile()); //ignores xbim and zip files
+            }
         }
 
         public static IStepFileHeader GetFileHeader(string fileName)
@@ -52,7 +53,7 @@ namespace Xbim.IO.Memory
                         if (entry == null) throw new FileLoadException($"File does not contain a valid model: {fileName}");
                         using (var reader = entry.Open())
                         {
-                            if(entry.Name.IsStepTextFile())
+                            if (entry.Name.IsStepTextFile())
                                 return GetStepFileHeader(reader, new MemoryModel(new Xbim.Ifc4.EntityFactory())); //using a dummy model to get the assembly correct
                             if (entry.Name.IsStepXmlFile())
                                 return XbimXmlReader4.ReadHeader(reader);
@@ -66,7 +67,7 @@ namespace Xbim.IO.Memory
                     throw new FileLoadException($"File is an invalid zip format: {fileName}");
                 }
             }
-           
+
             else if (fileName.IsStepXmlFile())
             {
                 using (var fileStream = File.OpenRead(fileName))
@@ -81,12 +82,12 @@ namespace Xbim.IO.Memory
         public static IStepFileHeader GetStepFileHeader(Stream stream, IModel model)
         {
             var parser = new XbimP21Parser(stream, null, -1);
-           
-            var stepHeader = new StepFileHeader(StepFileHeader.HeaderCreationMode.LeaveEmpty,model);
+
+            var stepHeader = new StepFileHeader(StepFileHeader.HeaderCreationMode.LeaveEmpty, model);
             parser.EntityCreate += (string name, long? label, bool header, out int[] ints) =>
             {
-                    //allow all attributes to be parsed
-                    ints = null;
+                //allow all attributes to be parsed
+                ints = null;
                 if (header)
                 {
                     switch (name)
@@ -102,7 +103,7 @@ namespace Xbim.IO.Memory
                     }
                 }
                 parser.Cancel = true; //done enough
-                    return null;
+                return null;
             };
             parser.Parse();
             return stepHeader;
@@ -123,7 +124,7 @@ namespace Xbim.IO.Memory
 
             _entityFactory = entityFactory;
             _instances = new EntityCollection(this, labelFrom);
-            Header = new StepFileHeader(StepFileHeader.HeaderCreationMode.InitWithXbimDefaults,this);
+            Header = new StepFileHeader(StepFileHeader.HeaderCreationMode.InitWithXbimDefaults, this);
             foreach (var schemasId in _instances.Factory.SchemasIds)
                 Header.FileSchema.Schemas.Add(schemasId);
             ModelFactors = new XbimModelFactors(Math.PI / 180, 1e-3, 1e-5);
@@ -171,10 +172,10 @@ namespace Xbim.IO.Memory
             return dropped;
         }
 
-        
+
         private void TryDrop(IPersistEntity entity, HashSet<IPersistEntity> dropped)
         {
-            
+
[…]
