### 2017-09-13 c9661aa
MemoryModel has a new function LoadStep21Part which takes partial file and loads it into current model. This is usefull for partial data exchange.
