+++ Xbim.IO.MemoryModel/Step21/Parser/StepP21Parser.cs
@@ -3,10 +3,10 @@
 // (see accompanying GPPGcopyright.rtf)
 
 // GPPG version 1.5.2
-// Machine:  DESKTOP-S5GEMU1
-// DateTime: 23/01/2017 10:53:26
-// UserName: Claudio
-// Input file <StepP21Parser.y - 24/10/2016 20:53:49>
+// Machine:  DESKTOP-VAJP4OB
+// DateTime: 13/09/2017 16:53:33
+// UserName: Martin
+// Input file <StepP21Parser.y - 13/09/2017 16:53:29>
 
 // options: lines gplex
 
@@ -53,11 +53,11 @@ public class ScanObj {
 [GeneratedCodeAttribute( "Gardens Point Parser Generator", "1.5.2")]
 public partial class P21Parser: ShiftReduceParser<ValueType, LexLocation>
 {
-  // Verbatim content from StepP21Parser.y - 24/10/2016 20:53:49
+  // Verbatim content from StepP21Parser.y - 13/09/2017 16:53:29
 #line 2 "StepP21Parser.y"
  public bool InHeader = false;
 #line default
-  // End verbatim content from StepP21Parser.y - 24/10/2016 20:53:49
+  // End verbatim content from StepP21Parser.y - 13/09/2017 16:53:29
 
 #pragma warning disable 649
   private static Dictionary<int, string> aliases;
@@ -268,138 +268,143 @@ public partial class P21Parser: ShiftReduceParser<ValueType, LexLocation>
     {InHeader=true; BeginHeader();}
 #line default
         break;
+      case 14: // stepFile -> model
+#line 59 "StepP21Parser.y"
+         { EndParse(); }
+#line default
+        break;
       case 15: // endSec -> ENDSEC
-#line 57 "StepP21Parser.y"
+#line 62 "StepP21Parser.y"
                {EndSec();}
 #line default
         break;
       case 16: // endSec -> ENDSEC, trailingSpace
-#line 58 "StepP21Parser.y"
+#line 63 "StepP21Parser.y"
                         {EndSec();}
 #line default
         break;
       case 19: // headerEntity -> entityType, listArgument, ';'
-#line 64 "StepP21Parser.y"
+#line 69 "StepP21Parser.y"
                                            {EndHeaderEntity();}
 #line default
         break;
       case 21: // endOfHeader -> DATA
-#line 68 "StepP21Parser.y"
+#line 73 "StepP21Parser.y"
  {  InHeader=false; EndHeader();  }
 #line default
         break;
       case 22: // argument -> IDENTITY
-#line 71 "StepP21Parser.y"
+#line 76 "StepP21Parser.y"
                             {SetObjectValue(CurrentSemanticValue.strVal);}
 #line default
         break;
       case 23: // argument -> INTEGER
-#line 72 "StepP21Parser.y"
+#line 77 "StepP21Parser.y"
                         {SetIntegerValue(CurrentSemanticValue.strVal);}
 #line default
         break;
       case 24: // argument -> FLOAT
-#line 73 "StepP21Parser.y"
+#line 78 "StepP21Parser.y"
                           {SetFloatValue(CurrentSemanticValue.strVal);}
 #line default
         break;
       case 25: // argument -> STRING
-#line 74 "StepP21Parser.y"
+#line 79 "StepP21Parser.y"
                        {SetStringValue(CurrentSemanticValue.strVal);}
 #line default
         break;
       case 26: // argument -> BOOLEAN
-#line 75 "StepP21Parser.y"
+#line 80 "StepP21Parser.y"
                         {SetBooleanValue(CurrentSemanticValue.strVal);}
 #line default
         break;
       case 27: // argument -> ENUM
-#line 76 "StepP21Parser.y"
+#line 81 "StepP21Parser.y"
[…]
