### 2017-09-13 8b4dc3d
Enhanced test (removed unnecessary code)
