### 2017-09-14 c10d05f
Changed interface for inverse relations caching, introduced new interface for entity caching for better controll over entity caching where applicable.
