### 2017-09-14 1148893
Updated to the latest Xbim.Common. Implemented changed and new interfaces for cache management.
