### 2017-09-14 c56562d
Updated to the latest Xbim.Common. Implemented changed and new interfaces for cache management.
