### 2017-09-15 0a21d4e
Changes made to schema version resolution so it doesn't throw any exceptions in the parser. MemoryModel can now open step21 text file with a single access or from the stream and it will resolve the schema during parsing operation.
