+++ Xbim.IO.MemoryModel/MemoryModel/MemoryModel.cs
@@ -32,7 +32,7 @@ namespace Xbim.IO.Memory
             }
         }
 
-        public static IStepFileHeader GetFileHeader(string fileName)
+        public static XbimSchemaVersion GetSchemaVersion(string fileName)
         {
             // need to get the header for each step file storage type
             //if it is a zip, xml or ifc text
@@ -40,7 +40,7 @@ namespace Xbim.IO.Memory
             {
                 using (var fileStream = File.OpenRead(fileName))
                 {
-                    return GetStepFileHeader(fileStream, new MemoryModel(new Xbim.Ifc4.EntityFactory())); //using a dummy model to get the assembly correct
+                    return GetStepFileXbimSchemaVersion(fileStream);
                 }
             }
             else if (fileName.IsStepZipFile())
@@ -54,9 +54,26 @@ namespace Xbim.IO.Memory
                         using (var reader = entry.Open())
                         {
                             if (entry.Name.IsStepTextFile())
-                                return GetStepFileHeader(reader, new MemoryModel(new Xbim.Ifc4.EntityFactory())); //using a dummy model to get the assembly correct
+                                return GetStepFileXbimSchemaVersion(reader);
                             if (entry.Name.IsStepXmlFile())
-                                return XbimXmlReader4.ReadHeader(reader);
+                            {
+                                XmlSchemaVersion schema;
+                                using (var xml = XmlReader.Create(reader))
+                                {
+                                    schema = XbimXmlReader4.ReadSchemaVersion(xml);
+                                }
+
+                                switch (schema)
+                                {
+                                    case XmlSchemaVersion.Ifc2x3:
+                                        return XbimSchemaVersion.Ifc2X3;
+                                    case XmlSchemaVersion.Ifc4Add1:
+                                    case XmlSchemaVersion.Ifc4:
+                                        return XbimSchemaVersion.Ifc4;
+                                    case XmlSchemaVersion.Unknown:
+                                        return XbimSchemaVersion.Unsupported;
+                                }
+                            }
                             else
                                 throw new FileLoadException($"File does not contain a valid model: {fileName}");
                         }
@@ -72,71 +89,167 @@ namespace Xbim.IO.Memory
             {
                 using (var fileStream = File.OpenRead(fileName))
                 {
-                    return XbimXmlReader4.ReadHeader(fileStream);
+                    XmlSchemaVersion schema;
+                    using (var reader = XmlReader.Create(fileStream))
+                    {
+                        schema = XbimXmlReader4.ReadSchemaVersion(reader);
+                    }
+
+                    switch (schema)
+                    {
+                        case XmlSchemaVersion.Ifc2x3:
+                            return XbimSchemaVersion.Ifc2X3;
+                        case XmlSchemaVersion.Ifc4Add1:
+                        case XmlSchemaVersion.Ifc4:
+                            return XbimSchemaVersion.Ifc4;
+                        case XmlSchemaVersion.Unknown:
+                            return XbimSchemaVersion.Unsupported;
+                    }
+
                 }
             }
-            else
-                throw new FileLoadException($"File is an invalid model format: {fileName}");
+
+            throw new FileLoadException($"File is an invalid model format: {fileName}");
         }
 
-        public static IStepFileHeader GetStepFileHeader(Stream stream, IModel model)
+        //public static IStepFileHeader GetStepFileHeader(Stream stream, IModel model)
+        //{
+        //    var parser = new XbimP21Parser(stream, null, -1);
+
[…]
