### 2017-09-20 fbbf9bc
ModelFactors exposed publicly
