### 2017-09-22 b7956ca
Performance tuning - Metadata
