+++ Xbim.IO.MemoryModel/MemoryModel/PartialParser.cs
@@ -7,14 +7,14 @@ namespace Xbim.IO.MemoryModel
 {
     internal class PartialParser : XbimP21Parser
     {
-        public PartialParser(string data, ExpressMetaData metadata): base(metadata)
+        public PartialParser(string data)
         {
             var scanner = new Scanner();
             scanner.SetSource(data, 0);
             Scanner = scanner;
         }
 
-        public PartialParser(Stream data, ExpressMetaData metadata) : base(data, metadata, 0)
+        public PartialParser(Stream data) : base(data, 0)
         {
         }
     }
