### 2017-09-26 748ad87
Further MemoryModel optimization
