+++ Xbim.IO.MemoryModel/Step21/Parser/OptimizedScanner.cs
@@ -1,0 +1,24 @@
+﻿namespace Xbim.IO.Optimized
+{
+    public sealed partial class Scanner
+    {
+        private XbimScanBuffer _buffer;
+
+        public void SetSource(XbimScanBuffer source)
+        {
+            _buffer = source;
+            buffer = source;
+            lNum = 0;
+            code = '\n'; // to initialize yyline, yycol and lineStart
+            GetCode();
+        }
+
+        public int yylabel
+        {
+            get
+            {
+                return _buffer.GetLabel(tokPos, tokEPos);
+            }
+        }
+    }
+}
