### 2017-09-26 fc97ed3
Further MemoryModel optimization
