+++ Xbim.IO.MemoryModel/Step21/XbimP21Parser.cs
@@ -53,8 +53,6 @@ namespace Xbim.IO.Step21
             get { return ListNestLevel > 0 ? _nestedIndex.ToArray() : null; }
         }
 
-        protected readonly ExpressMetaData Metadata;
-
         public XbimP21Parser(Stream strm, long streamSize)
             : base(strm)
         {
@@ -63,11 +61,11 @@ namespace Xbim.IO.Step21
             if (streamSize > 0)
             {
                 _streamSize = streamSize;
-                entityApproxCount = Convert.ToInt32(_streamSize/50); //average 50 bytes per entity.
+                entityApproxCount = Convert.ToInt32(_streamSize / 50); //average 50 bytes per entity.
             }
 
             _entities = new Dictionary<long, IPersist>(entityApproxCount);
-            _deferredReferences = new List<DeferredReference>(entityApproxCount/2); //assume 50% deferred
+            _deferredReferences = new List<DeferredReference>(entityApproxCount / 2); //assume 50% deferred
             ErrorCount = 0;
         }
 
@@ -76,7 +74,7 @@ namespace Xbim.IO.Step21
             //Metadata = metadata;
             const int entityApproxCount = 5000;
             _entities = new Dictionary<long, IPersist>(entityApproxCount);
-            _deferredReferences = new List<DeferredReference>(entityApproxCount/2); //assume 50% deferred
+            _deferredReferences = new List<DeferredReference>(entityApproxCount / 2); //assume 50% deferred
             ErrorCount = 0;
         }
 
@@ -182,9 +180,9 @@ namespace Xbim.IO.Step21
             _processStack.Push(CurrentInstance);
             if (_streamSize == -1 || ProgressStatus == null) return;
 
-            var sc = (Scanner) Scanner;
+            var sc = (Scanner)Scanner;
             double pos = sc.Buffer.Pos;
-            var newPercentage = Convert.ToInt32(pos/_streamSize*100.0);
+            var newPercentage = Convert.ToInt32(pos / _streamSize * 100.0);
             if (newPercentage <= _percentageParsed) return;
             _percentageParsed = newPercentage;
             ProgressStatus(_percentageParsed, "Parsing");
@@ -322,7 +320,7 @@ namespace Xbim.IO.Step21
                 if (CurrentInstance.Entity != null)
                     CurrentInstance.Entity.Parse(CurrentInstance.CurrentParamIndex, PropertyValue, NestedIndex);
             }
-            catch (Exception )
+            catch (Exception)
             {
                 if (ErrorCount > MaxErrorCount)
                     throw new XbimParserException("Too many errors in file, parser execution terminated");
@@ -330,11 +328,9 @@ namespace Xbim.IO.Step21
                 var mainEntity = _processStack.Last();
                 if (mainEntity != null)
                 {
-                    var expressType = Metadata.ExpressType(mainEntity.Entity);
-                    Logger?.LogError("Entity #{0,-5} {1}, error at parameter {2}-{3} value = {4}",
+                    Logger?.LogError("Entity #{0,-5} {1}, error at parameter {2} value = {4}",
                                                mainEntity.EntityLabel, mainEntity.Entity.GetType().Name.ToUpper(),
                                                mainEntity.CurrentParamIndex + 1,
-                                               expressType.Properties[mainEntity.CurrentParamIndex + 1].PropertyInfo.Name,
                                                value);
                 }
                 else
@@ -367,7 +363,7 @@ namespace Xbim.IO.Step21
                 if (CurrentInstance.Entity != null)
                     CurrentInstance.Entity.Parse(CurrentInstance.CurrentParamIndex, PropertyValue, NestedIndex);
             }
-            catch (Exception )
+            catch (Exception)
[…]
