### 2017-09-26 60d679a
Purge WIP code. This implementation is 40% faster than the original one.
