### 2017-09-26 04dd230
Purge WIP code. This implementation is 40% faster than the original one.
