+++ Xbim.IO.MemoryModel/Step21/Parser/StepP21Scanner.cs
@@ -1,2923 +1,0 @@
-//
-//  This CSharp output file generated by Gardens Point LEX
-//  Gardens Point LEX (GPLEX) is Copyright (c) John Gough, QUT 2006-2014.
-//  Output produced by GPLEX is the property of the user.
-//  See accompanying file GPLEXcopyright.rtf.
-//
-//  GPLEX Version:  1.2.2
-//  Machine:  DESKTOP-VAJP4OB
-//  DateTime: 26/09/2017 10:21:53
-//  UserName: Martin
-//  GPLEX input file <StepP21Scanner.LEX - 26/09/2017 10:21:49>
-//  GPLEX frame file <embedded resource>
-//
-//  Option settings: verbose, parser, minimize
-//  Option settings: noCompressNext, noPersistBuffer, embedbuffers
-//
-
-//
-// Revised backup code
-// Version 1.2.1 of 24-June-2013
-//
-//
-#define BACKUP
-#define BYTEMODE
-
-using System;
-using System.IO;
-using System.Text;
-using System.Globalization;
-using System.Collections.Generic;
-using System.Runtime.Serialization;
-using System.Diagnostics.CodeAnalysis;
-
-using Xbim.IO.Parser;
-
-namespace Xbim.IO.Optimized
-{   
-    /// <summary>
-    /// Summary Canonical example of GPLEX automaton
-    /// </summary>
-    
-#if STANDALONE
-    //
-    // These are the dummy declarations for stand-alone GPLEX applications
-    // normally these declarations would come from the parser.
-    // If you declare /noparser, or %option noparser then you get this.
-    //
-
-     public enum Tokens
-    { 
-      EOF = 0, maxParseToken = int.MaxValue 
-      // must have at least these two, values are almost arbitrary
-    }
-
-     public abstract class ScanBase
-    {
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "yylex")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "yylex")]
-        public abstract int yylex();
-
-        [SuppressMessage("Microsoft.Naming", "CA1709:IdentifiersShouldBeCasedCorrectly", MessageId = "yywrap")]
-        [SuppressMessage("Microsoft.Naming", "CA1704:IdentifiersShouldBeSpelledCorrectly", MessageId = "yywrap")]
-        protected virtual bool yywrap() { return true; }
-
-#if BABEL
-        protected abstract int CurrentSc { get; set; }
-        // EolState is the 32-bit of state data persisted at 
-        // the end of each line for Visual Studio colorization.  
-        // The default is to return CurrentSc.  You must override
-        // this if you want more complicated behavior.
[…]
