### 2017-09-26 c7c6b71
Loading partial files uses scanner instead of parser as well. This is 40% faster.
