### 2017-09-27 7b651a9
Errors tolerance for loading "bad" files is increased
