### 2017-09-28 ceb556f
Added test case for Issue #107.
Passes with Memory, but not with Esent.
