+++ Xbim.IO/Step21/XbimP21Parser.cs
@@ -41,7 +41,7 @@ namespace Xbim.IO.Step21
         protected PropertyValue PropertyValue;
         private readonly List<DeferredReference> _deferredReferences;
         private readonly double _streamSize = -1;
-        public static int MaxErrorCount = 100;
+        public static int MaxErrorCount = 200;
         private int _percentageParsed;
         private bool _deferListItems;
         public bool Cancel = false;
