### 2017-09-30 #395 @CBenghi
Hello @arfon,

I've looked again at the list of editors and (looking at all individual github pages) you seem to be short on .Net Platform reviewers.
I might be wrong, but I expect a resurgence of .Net now that the new cross platform core 2.0 is out.

If I manage to find a volunteers with an established .Net background (e.g. .net foundation or otherwise active in the open-source net ecosystem) would you accept their review?

Should they have academic research background?
Thanks,
Claudio
