### 2017-10-04 #395 @arfon
> I've looked again at the list of editors and (looking at all individual github pages) you seem to be short on .Net Platform reviewers.

Yes, that's definitely true! Feel free to suggest some potential reviewers and we'll see what we can do.
