### 2017-10-09 5c1eb74
Interface extended IXbimGeometryEngine.CreateSolidSet for IIfcSweptAreaSolid to deal with compound profile definitions
