### 2017-10-12 bd3619a
Minor changes in the parser to increase its resiliance. Because it doesn't have any checks for the input structure it should try to catch up again at the beginning of next entity should it fail.
