+++ Xbim.IO.MemoryModel/Step21/XbimP21Scanner.cs
@@ -31,6 +31,13 @@ using Xbim.IO.Step21.Parser;
 
 namespace Xbim.IO.Step21
 {
+    /// <summary>
+    /// This class is meant to replace original parser which was generated using GPPG.
+    /// Actual implementation of the original parser was actually only scanner with little
+    /// added value but it had significant overhead on processing spead. This implementation
+    /// uses the same methods but binds them to the scanner directly. This cuts off part of the
+    /// structure security but it also cuts off processing time by about 20%.
+    /// </summary>
     public class XbimP21Scanner
     {
         private ILogger _logger;
@@ -48,6 +55,8 @@ namespace Xbim.IO.Step21
         private int _percentageParsed;
         private bool _deferListItems;
         private readonly List<int> _nestedIndex = new List<int>();
+
+        public bool Cancel = false;
 
         public int[] NestedIndex
         {
@@ -83,103 +92,139 @@ namespace Xbim.IO.Step21
             _deferredReferences = new List<DeferredReference>(entityApproxCount / 2); //assume 50% deferred
         }
 
-        public void Parse()
-        { 
+        public bool Parse()
+        {
+            var eofToken = (int)Tokens.EOF;
             var tok = _scanner.yylex();
-            while (tok != (int)Tokens.EOF)
+            while (tok != eofToken && !Cancel)
             {
-                if (tok >= 63)
+                try
                 {
-                    Tokens t = (Tokens)tok;
-                    switch (t)
+                    if (tok >= 63)
                     {
-                        case Tokens.HEADER:
-                            BeginHeader();
-                            break;
-                        case Tokens.ENDSEC:
-                            EndSec();
-                            break;
-                        case Tokens.DATA:
-                            BeginData();
-                            break;
-                        case Tokens.ENTITY:
-                            NewEntity(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.TYPE:
-                            SetType(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.INTEGER:
-                            SetIntegerValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.FLOAT:
-                            SetFloatValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.STRING:
-                            SetStringValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.BOOLEAN:
-                            SetBooleanValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.IDENTITY:
-                            SetObjectValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.HEXA:
-                            SetHexValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.ENUM:
-                            SetEnumValue(_scanner.yylval.strVal);
-                            break;
-                        case Tokens.NONDEF:
-                            SetNonDefinedValue();
-                            break;
-                        case Tokens.OVERRIDE:
-                            SetOverrideValue();
-                            break;
+                        Tokens t = (Tokens)tok;
+                        switch (t)
+                        {
+                            case Tokens.HEADER:
+                                BeginHeader();
+                                break;
+                            case Tokens.ENDSEC:
+                                EndSec();
+                                break;
+                            case Tokens.DATA:
+                                BeginData();
+                                break;
+                            case Tokens.ENTITY:
+                                NewEntity(_scanner.yylval.strVal);
+                                break;
+                            case Tokens.TYPE:
+                                SetType(_scanner.yylval.strVal);
+                                break;
+                            case Tokens.INTEGER:
+                                SetIntegerValue(_scanner.yylval.strVal);
+                                break;
+                            case Tokens.FLOAT:
+                                SetFloatValue(_scanner.yylval.strVal);
[…]
