### 2017-10-12 2b6df8f
Generall exception also checked for maximum number of errors.
