### 2017-10-12 29acd4d
Code clean up.
