### 2017-10-12 1889113
Code clean up.
