+++ Xbim.IO.MemoryModel.Tests/PartialFileTests.cs
@@ -32,7 +32,8 @@ namespace Xbim.IfcCore.UnitTests
                         insModel.InsertCopy(extrusion, new Common.XbimInstanceHandleMap(mm, insModel), null, false, true, true);
                         using (var partModel = new MemoryModel(new EntityFactory()))
                         {
-                            partModel.LoadStep21Part(part);
+                            var errs = partModel.LoadStep21Part(part);
+                            Assert.IsTrue(errs == 0);
                             Assert.IsTrue(insModel.Instances.Count == partModel.Instances.Count);
 
                             var ext2 = partModel.Instances.FirstOrDefault<IfcExtrudedAreaSolid>();
