### 2017-10-12 772d75e
Progress reporting improved.
