### 2017-10-24 2d52e06
Ability to define names of types to skip. This allows to minimize memory footprint and further speed up model parsing.
