### 2017-10-24 3b94877
Ability to define names of types to skip. This allows to minimize memory footprint and further speed up model parsing.
