### 2017-10-30 2d2d7d9
Provided a method to gauge availability of geometry without blocking the Esent file.
