### 2017-10-30 #395 @CBenghi
@arfon,
I've tried with no luck to get a suitable reviewer, but I'm not finding it easy.

I'll propose a different path.
I can take you (or your preferred editor) through the hurdles of the .Net technology for the review of Xbim.Essentials and thereafter I'll take care of any other .Net reviews that might be submitted.

All it takes is a Windows machine, and a little time. Visual Studio is free and installs relatively fast in the latest version. All other dependencies needed are downloaded automatically once you have the code downloaded from Github.

I'm quite keen on this, I'd be happy to support the review with custom videos or any questions you might have along the way.

Would this be possible?
