### 2017-10-31 #395 @arfon
> All it takes is a Windows machine, and a little time. Visual Studio is free and installs relatively fast in the latest version. All other dependencies needed are downloaded automatically once you have the code downloaded from Github.

I wouldn't be able to do this sorry. Would any of the @openjournals/joss-editors be able to do this?
