### 2017-11-01 b1c80d9
Minor change of member visibility in parser to make it consistent with other parts of the code.
