### 2017-11-01 722c7e7
Reading COBieExpress from spreadsheet fixed to assign levels to facility explicitly because COBie XLSX standard only contains this relation implicitly.
