### 2017-11-01 ab6f106
Changes made in Xbim.Tesselator to write results as floats, not as doubles. This is for backwards binary compatibility with older files but keeps double precission as introduced by François-Olivier Mayette for improved geometry processing.
