### 2017-11-01 965190c
Reversed back changes in binary reader made by François-Olivier Mayette in PR #103. This should make xBIM binary compatible with older files. Double precision is kept in Xbim.Tesselator for precision of data transformations. It is only converted to single when serialized into binary.
