### 2017-11-14 da3b953
Fix for CodePageOverride
Some apps creating IFC-files are encoding one-byte-char strings with implicitly specified encoding using CodePages other then ISO8859-1. That is a violation of the IFC specification, but such files exist, and the IfcStore class users should have a way to override an implicit encoding ehen parsing Step21-files.
The fix provides such a way via an optional argument:
IfcStore.Open(..., int codePageOverride = -1)
The fix is implemented only for a Esent underlying model.
An example file and tests are included.
