+++ Xbim.Ifc/IfcStore.cs
@@ -125,14 +125,22 @@ namespace Xbim.Ifc
             if (EntityModified != null) EntityModified.Invoke(entity, property);
         }
 
-        private static EsentModel CreateEsentModel(IfcSchemaVersion schema)
+        private static EsentModel CreateEsentModel(IfcSchemaVersion schema, int codePageOverride)
         {
             switch (schema)
             {
                 case IfcSchemaVersion.Ifc4:
-                    return new EsentModel(new Ifc4.EntityFactory());
+                {
+                    var model = new EsentModel(new Ifc4.EntityFactory());
+                    model.CodePageOverride = codePageOverride;
+                    return model;
+                }
                 case IfcSchemaVersion.Ifc2X3:
-                    return new EsentModel(new Ifc2x3.EntityFactory());
+                {
+                    var model = new EsentModel(new Ifc2x3.EntityFactory());
+                    model.CodePageOverride = codePageOverride;
+                    return model;
+                }
                 default:
                     throw new NotSupportedException("IfcStore only supports IFC schemas");
             }
@@ -165,7 +173,7 @@ namespace Xbim.Ifc
         /// <param name="accessMode">Access mode to the stream. This is only important if you choose EsentModel. MemoryModel is completely in memory so this is not relevant</param>
         /// <param name="progDelegate">Progress reporting delegate</param>
         /// <returns></returns>
-        public static IfcStore Open(Stream data, IfcStorageType dataType, IfcSchemaVersion schema, XbimModelType modelType, XbimEditorCredentials editorDetails = null, XbimDBAccess accessMode = XbimDBAccess.Read, ReportProgressDelegate progDelegate = null)
+        public static IfcStore Open(Stream data, IfcStorageType dataType, IfcSchemaVersion schema, XbimModelType modelType, XbimEditorCredentials editorDetails = null, XbimDBAccess accessMode = XbimDBAccess.Read, ReportProgressDelegate progDelegate = null, int codepageOverride = -1)
         {
             //any Esent model needs to run from the file so we need to create a temporal one
             var xbimFilePath = Path.GetTempFileName();
@@ -198,14 +206,14 @@ namespace Xbim.Ifc
                         }
                     }
                     {
-                        var model = CreateEsentModel(schema);
+                        var model = CreateEsentModel(schema, codepageOverride);
                         model.Open(xbimFilePath, accessMode, progDelegate);
                         return new IfcStore(model, schema, editorDetails, xbimFilePath);
                     }
                 case IfcStorageType.IfcXml:
                     if (modelType == XbimModelType.EsentModel)
                     {
-                        var model = CreateEsentModel(schema);
+                        var model = CreateEsentModel(schema, codepageOverride);
                         if (model.CreateFrom(data, data.Length, dataType, xbimFilePath, progDelegate, true, true))
                             return new IfcStore(model, schema, editorDetails, xbimFilePath);
                         else
@@ -222,7 +230,7 @@ namespace Xbim.Ifc
                 case IfcStorageType.Ifc:
                     if (modelType == XbimModelType.EsentModel)
                     {
-                        var model = CreateEsentModel(schema);
+                        var model = CreateEsentModel(schema, codepageOverride);
                         if (model.CreateFrom(data, data.Length, dataType, xbimFilePath, progDelegate, true, true))
                             return new IfcStore(model, schema, editorDetails, xbimFilePath);
                         else
@@ -240,7 +248,7 @@ namespace Xbim.Ifc
                 case IfcStorageType.Zip:
                     if (modelType == XbimModelType.EsentModel)
                     {
[…]
