### 2017-11-20 cd896a5
Fix in TableStore. ReferenceContext signalized there were data even if there were none in the row.
