+++ Xbim.IO.Tests/TableStorageTests.cs
@@ -24,6 +24,7 @@ namespace Xbim.IO.Tests
         public void ContactsImport()
         {
             const string file = @"c:\Users\Martin\Source\Samples\cutdown.xlsx";
+            string report;
             using (var model = CobieModel.ImportFromTable(file, out report))
             {
                 var contacts = model.Instances.OfType<CobieContact>().ToList();
