### 2017-11-22 1143d6c
If the row doesn't have a key it is not imported.
