+++ Xbim.IO.TableStore/ReferenceContext.cs
@@ -77,7 +77,8 @@ namespace Xbim.IO.TableStore
 
         public IEnumerable<ReferenceContext> AllChildren
         {
-            get {
+            get
+            {
                 foreach (var child in Children)
                 {
                     yield return child;
@@ -87,7 +88,7 @@ namespace Xbim.IO.TableStore
                     }
                 }
             }
-        } 
+        }
 
         public bool HasEnumerationOnPath
         {
@@ -109,15 +110,35 @@ namespace Xbim.IO.TableStore
             }
         }
 
-        public bool IsReference { get
+        public bool IsReference
         {
-            return (Mapping != null && Mapping.Status == DataStatus.Reference) || ScalarChildren.Any(c => c.Mapping != null && c.Mapping.Status == DataStatus.Reference);
-        } }
+            get
+            {
+                return (Mapping != null && Mapping.Status == DataStatus.Reference) || ScalarChildren.Any(c => c.Mapping != null && c.Mapping.Status == DataStatus.Reference);
+            }
+        }
 
         /// <summary>
         /// Any scalar child of any children has values loaded from a row
         /// </summary>
         public bool HasData { get; private set; }
+
+        /// <summary>
+        /// Any scalar identity data exist
+        /// </summary>
+        public bool HasKeyData { get; private set; }
+
+        /// <summary>
+        /// There is at least one identity requirement in the mappinge
+        /// </summary>
+        public bool HasKeyRequirements
+        {
+            get
+            {
+                return ScalarChildren.Any(c => c.Mapping.IsKey);
+            }
+        }
+
 
         public ReferenceContext(TableStore store, ClassMapping cMapping)
         {
@@ -130,7 +151,7 @@ namespace Xbim.IO.TableStore
             if (cMapping.PropertyMappings == null || !cMapping.PropertyMappings.Any())
                 return;
 
-            var parentPath = ("parent." + cMapping.ParentPath) .Split(new [] {'.'}, StringSplitOptions.RemoveEmptyEntries).ToList();
+            var parentPath = ("parent." + cMapping.ParentPath).Split(new[] { '.' }, StringSplitOptions.RemoveEmptyEntries).ToList();
             foreach (var mapping in cMapping.PropertyMappings)
             {
                 var path = mapping.Paths.FirstOrDefault();
@@ -155,7 +176,7 @@ namespace Xbim.IO.TableStore
             //add special mapping for a parent path
             if (!cMapping.IsRoot)
             {
-                var pMap = new PropertyMapping{_Paths = string.Join(".", parentPath), Status = DataStatus.Reference};
+                var pMap = new PropertyMapping { _Paths = string.Join(".", parentPath), Status = DataStatus.Reference };
                 AddMapping(pMap, parentPath);
             }
 
@@ -205,7 +226,7 @@ namespace Xbim.IO.TableStore
 
             PropertyTypeHint = Store.MetaData.ExpressType(MetaProperty != null ?
                     MetaProperty.EnumerableType ?? MetaProperty.PropertyInfo.PropertyType :
-                    (PropertyInfo != null ? PropertyInfo.PropertyType : null) 
+                    (PropertyInfo != null ? PropertyInfo.PropertyType : null)
                 );
 
 
@@ -213,8 +234,8 @@ namespace Xbim.IO.TableStore
             var isEnumerable = MetaProperty != null && MetaProperty.EnumerableType != null;
             if (isEnumerable)
             {
-                if(MetaProperty.EnumerableType.IsValueType || 
-                    MetaProperty.EnumerableType == typeof(string) || 
+                if (MetaProperty.EnumerableType.IsValueType ||
+                    MetaProperty.EnumerableType == typeof(string) ||
                     typeof(IExpressValueType).IsAssignableFrom(MetaProperty.EnumerableType))
[…]
