### 2017-11-22 08db8f3
Create CONTRIBUTING.md
It's good practice to have such file.
