### 2017-11-27 19e3e70
Links to pull request documentation added as many people don't know what pull request is.
