### 2017-11-27 #395 @Haacked
Hey @arfon! It's been a long time. I've been asked to review this project as a subject matter expert in .NET and I was happy to see your involvement.

I'm happy to help review this. Just let me know next steps.
