### 2017-11-27 2e9c1d2
Link the license name to the license file
The README mentions a "license" folder, but I couldn't find one. I did find the actual license file, so it makes sense to simply link to that file.
