### 2017-11-28 87a55a3
Fixed failing test case "CanParseNewlinesInStrings" by cherrypicking improved lexer from develop.
