+++ Xbim.IO/Step21/Parser/StepP21Lex.cs
@@ -5,10 +5,10 @@
 //  See accompanying file GPLEXcopyright.rtf.
 //
 //  GPLEX Version:  1.2.2
-//  Machine:  DESKTOP-S5GEMU1
-//  DateTime: 26/09/2016 22:38:48
-//  UserName: Claudio
-//  GPLEX input file <StepP21Lex.LEX - 26/09/2016 22:38:44>
+//  Machine:  C10200216
+//  DateTime: 28/11/2017 16:02:30
+//  UserName: sgmk2
+//  GPLEX input file <StepP21Lex.LEX - 28/11/2017 16:01:53>
 //  GPLEX frame file <embedded resource>
 //
 //  Option settings: verbose, parser, minimize
@@ -123,8 +123,8 @@ namespace Xbim.IO.Parser
         
         enum Result {accept, noMatch, contextFound};
 
-        const int maxAccept = 88;
-        const int initial = 89;
+        const int maxAccept = 83;
+        const int initial = 84;
         const int eofNum = 0;
         const int goStart = -1;
         const int INITIAL = 0;
@@ -162,65 +162,69 @@ public static int Pass = 1;
 #region ScannerTables
     struct Table {
         public int min; public int rng; public int dflt;
-        public short[] nxt;
-        public Table(int m, int x, int d, short[] n) {
+        public sbyte[] nxt;
+        public Table(int m, int x, int d, sbyte[] n) {
             min = m; rng = x; dflt = d; nxt = n;
         }
     };
 
-    static int[] startState = new int[] {89, 129, 0};
+    static int[] startState = new int[] {84, 108, 0};
 
-    static Table[] NxS = new Table[131] {
+    static Table[] NxS = new Table[110] {
 /* NxS[   0] */ new Table(0, 0, 0, null), // Shortest string ""
 /* NxS[   1] */ // Shortest string ""
-      new Table(0, 1, -1, new short[] {1}),
+      new Table(0, 1, -1, new sbyte[] {1}),
 /* NxS[   2] */ new Table(0, 0, -1, null), // Shortest string "\x01"
 /* NxS[   3] */ new Table(0, 0, -1, null), // Shortest string "\t"
 /* NxS[   4] */ new Table(0, 0, -1, null), // Shortest string "\n"
 /* NxS[   5] */ new Table(0, 0, -1, null), // Shortest string "\r"
 /* NxS[   6] */ new Table(0, 0, -1, null), // Shortest string "\x20"
 /* NxS[   7] */ // Shortest string "!"
-      new Table(48, 75, -1, new short[] {87, 87, 87, 87, 87, 87, 
-          87, 87, 87, 87, -1, -1, -1, -1, -1, -1, -1, 87, 87, 87, 87, 87, 
-          87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 
-          87, 87, 87, 87, 87, -1, -1, -1, -1, 87, -1, 87, 87, 87, 87, 87, 
-          87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 87, 
-          87, 87, 87, 87, 87}),
[…]
