+++ Xbim.IO/Step21/Parser/StepP21Parser.cs
@@ -3,10 +3,10 @@
 // (see accompanying GPPGcopyright.rtf)
 
 // GPPG version 1.5.2
-// Machine:  DESKTOP-S5GEMU1
-// DateTime: 26/09/2016 22:38:48
-// UserName: Claudio
-// Input file <StepP21Parser.y - 29/06/2016 23:31:29>
+// Machine:  C10200216
+// DateTime: 28/11/2017 16:02:31
+// UserName: sgmk2
+// Input file <StepP21Parser.y - 13/10/2016 17:38:42>
 
 // options: lines gplex
 
@@ -53,11 +53,11 @@ public class ScanObj {
 [GeneratedCodeAttribute( "Gardens Point Parser Generator", "1.5.2")]
 public partial class P21Parser: ShiftReduceParser<ValueType, LexLocation>
 {
-  // Verbatim content from StepP21Parser.y - 29/06/2016 23:31:29
+  // Verbatim content from StepP21Parser.y - 13/10/2016 17:38:42
 #line 2 "StepP21Parser.y"
  public bool InHeader = false;
 #line default
-  // End verbatim content from StepP21Parser.y - 29/06/2016 23:31:29
+  // End verbatim content from StepP21Parser.y - 13/10/2016 17:38:42
 
 #pragma warning disable 649
   private static Dictionary<int, string> aliases;
