+++ Xbim.IO.Tests/ParsingTests.cs
@@ -147,25 +147,14 @@ namespace Xbim.MemoryModel.Tests
 
         [TestMethod]
         [DeploymentItem("TestFiles\\Issue107.zip")]
-        public void Issue107()
+        public void Issue107OnMemoryModel()
         {
-            // a merged PR on issue 107 makes the memory model more tolerant of duplicated items.
-            //
+            // a merged PR on issue 107 makes the memory model more tolerant of bad files.
+            // an equivalent test for esent is available
             using (var model = new Xbim.IO.Memory.MemoryModel(new Ifc2x3.EntityFactory()))
             {
                 var errCount = model.LoadZip("Issue107.zip");
                 Assert.AreEqual(140, errCount);
-            }
-
-            // the same does not apply to the database version though.
-            // an error is thrown in 
-            // Xbim.IO\Esent\EsentEntityCursor.cs
-            // at
-            // internal void AddEntity(int currentLabel, short typeId, IEnumerable<int> indexKeys, byte[] data, bool? indexed, EsentLazyDBTransaction? trans = null)
-            // because the label already exists and esent does not tolerate that nicely.
-            using (var model = new Xbim.IO.Esent.EsentModel(new Ifc2x3.EntityFactory()))
-            {
-                var errCount = model.CreateFrom("Issue107.zip"); 
             }
         }
 
