### 2017-11-29 #473 @whedon
[REVIEW]: Xbim.Essentials: a library for interoperable building information applications
**Submitting author:** @CBenghi (<a href="http://orcid.org/0000-0003-3252-5461">Claudio Benghi</a>)
**Repository:** <a href="https://github.com/xBimTeam/XbimEssentials" target ="_blank">https://github.com/xBimTeam/XbimEssentials</a>
**Version:** 4.0
