### 2017-11-29 #473 @arfon
@Haacked - please work through the checklist at the top of this issue. If you have any questions, don't hesitate to ask.

We also have some [reviewer guidelines](http://joss.theoj.org/about#reviewer_guidelines) which should help you understand the purpose and scope of the review.
