### 2017-11-30 2791c27
Control regeneration after a while.
