### 2017-11-30 3a41051
Upgrade to IFC4Add2 (we were implementing IFC4Add1). Step 21 implemented, XML definitions need to be updated and upgraded as well
