+++ Xbim.Ifc4/GeometryResource/IfcToroidalSurface.cs
@@ -1,0 +1,142 @@
+// ------------------------------------------------------------------------------
+// <auto-generated>
+//     This code was generated by a tool Xbim.CodeGeneration 
+//  
+//     Changes to this file may cause incorrect behaviour and will be lost if
+//     the code is regenerated.
+// </auto-generated>
+// ------------------------------------------------------------------------------
+
+using Xbim.Ifc4.MeasureResource;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using Xbim.Common;
+using Xbim.Common.Exceptions;
+using Xbim.Ifc4.Interfaces;
+using Xbim.Ifc4.GeometryResource;
+//## Custom using statements
+//##
+
+namespace Xbim.Ifc4.Interfaces
+{
+	/// <summary>
+    /// Readonly interface for IfcToroidalSurface
+    /// </summary>
+	// ReSharper disable once PartialTypeWithSinglePart
+	public partial interface @IIfcToroidalSurface : IIfcElementarySurface
+	{
+		IfcPositiveLengthMeasure @MajorRadius { get;  set; }
+		IfcPositiveLengthMeasure @MinorRadius { get;  set; }
+	
+	}
+}
+
+namespace Xbim.Ifc4.GeometryResource
+{
+	[ExpressType("IfcToroidalSurface", 1328)]
+	// ReSharper disable once PartialTypeWithSinglePart
+	public  partial class @IfcToroidalSurface : IfcElementarySurface, IInstantiableEntity, IIfcToroidalSurface, IContainsEntityReferences, IEquatable<@IfcToroidalSurface>
+	{
+		#region IIfcToroidalSurface explicit implementation
+		IfcPositiveLengthMeasure IIfcToroidalSurface.MajorRadius { 
+ 
+			get { return @MajorRadius; } 
+			set { MajorRadius = value;}
+		}	
+		IfcPositiveLengthMeasure IIfcToroidalSurface.MinorRadius { 
+ 
+			get { return @MinorRadius; } 
+			set { MinorRadius = value;}
+		}	
+		 
+		#endregion
+
+		//internal constructor makes sure that objects are not created outside of the model/ assembly controlled area
+		internal IfcToroidalSurface(IModel model, int label, bool activated) : base(model, label, activated)  
+		{
+		}
+
+		#region Explicit attribute fields
+		private IfcPositiveLengthMeasure _majorRadius;
+		private IfcPositiveLengthMeasure _minorRadius;
+		#endregion
+	
+		#region Explicit attribute properties
+		[EntityAttribute(2, EntityAttributeState.Mandatory, EntityAttributeType.None, EntityAttributeType.None, -1, -1, 4)]
+		public IfcPositiveLengthMeasure @MajorRadius 
+		{ 
+			get 
+			{
+				if(_activated) return _majorRadius;
+				Activate();
+				return _majorRadius;
+			} 
+			set
+			{
+				SetValue( v =>  _majorRadius = v, _majorRadius, value,  "MajorRadius", 2);
+			} 
+		}	
+		[EntityAttribute(3, EntityAttributeState.Mandatory, EntityAttributeType.None, EntityAttributeType.None, -1, -1, 5)]
+		public IfcPositiveLengthMeasure @MinorRadius 
+		{ 
+			get 
+			{
+				if(_activated) return _minorRadius;
+				Activate();
+				return _minorRadius;
+			} 
+			set
+			{
+				SetValue( v =>  _minorRadius = v, _minorRadius, value,  "MinorRadius", 3);
+			} 
+		}	
+		#endregion
+
+
+
+
+		#region IPersist implementation
+		public override void Parse(int propIndex, IPropertyValue value, int[] nestedIndex)
+		{
+			switch (propIndex)
+			{
[…]
