+++ Xbim.Ifc4/EntityFactory.cs
@@ -363,9 +363,12 @@ namespace Xbim.Ifc4
 				case "IFCIMAGETEXTURE": return new IfcImageTexture ( model, entityLabel, activated );
 				case "IFCINDEXEDCOLOURMAP": return new IfcIndexedColourMap ( model, entityLabel, activated );
 				case "IFCINDEXEDPOLYCURVE": return new IfcIndexedPolyCurve ( model, entityLabel, activated );
+				case "IFCINDEXEDPOLYGONALFACE": return new IfcIndexedPolygonalFace ( model, entityLabel, activated );
+				case "IFCINDEXEDPOLYGONALFACEWITHVOIDS": return new IfcIndexedPolygonalFaceWithVoids ( model, entityLabel, activated );
 				case "IFCINDEXEDTRIANGLETEXTUREMAP": return new IfcIndexedTriangleTextureMap ( model, entityLabel, activated );
 				case "IFCINTERCEPTOR": return new IfcInterceptor ( model, entityLabel, activated );
 				case "IFCINTERCEPTORTYPE": return new IfcInterceptorType ( model, entityLabel, activated );
+				case "IFCINTERSECTIONCURVE": return new IfcIntersectionCurve ( model, entityLabel, activated );
 				case "IFCINVENTORY": return new IfcInventory ( model, entityLabel, activated );
 				case "IFCIRREGULARTIMESERIES": return new IfcIrregularTimeSeries ( model, entityLabel, activated );
 				case "IFCIRREGULARTIMESERIESVALUE": return new IfcIrregularTimeSeriesValue ( model, entityLabel, activated );
@@ -462,6 +465,7 @@ namespace Xbim.Ifc4
 				case "IFCPOINTONSURFACE": return new IfcPointOnSurface ( model, entityLabel, activated );
 				case "IFCPOLYLOOP": return new IfcPolyLoop ( model, entityLabel, activated );
 				case "IFCPOLYGONALBOUNDEDHALFSPACE": return new IfcPolygonalBoundedHalfSpace ( model, entityLabel, activated );
+				case "IFCPOLYGONALFACESET": return new IfcPolygonalFaceSet ( model, entityLabel, activated );
 				case "IFCPOLYLINE": return new IfcPolyline ( model, entityLabel, activated );
 				case "IFCPOSTALADDRESS": return new IfcPostalAddress ( model, entityLabel, activated );
 				case "IFCPRESENTATIONLAYERASSIGNMENT": return new IfcPresentationLayerAssignment ( model, entityLabel, activated );
@@ -578,6 +582,7 @@ namespace Xbim.Ifc4
 				case "IFCSIUNIT": return new IfcSIUnit ( model, entityLabel, activated );
 				case "IFCSANITARYTERMINAL": return new IfcSanitaryTerminal ( model, entityLabel, activated );
 				case "IFCSANITARYTERMINALTYPE": return new IfcSanitaryTerminalType ( model, entityLabel, activated );
+				case "IFCSEAMCURVE": return new IfcSeamCurve ( model, entityLabel, activated );
 				case "IFCSECTIONPROPERTIES": return new IfcSectionProperties ( model, entityLabel, activated );
 				case "IFCSECTIONREINFORCEMENTPROPERTIES": return new IfcSectionReinforcementProperties ( model, entityLabel, activated );
 				case "IFCSECTIONEDSPINE": return new IfcSectionedSpine ( model, entityLabel, activated );
@@ -604,6 +609,7 @@ namespace Xbim.Ifc4
 				case "IFCSPATIALZONE": return new IfcSpatialZone ( model, entityLabel, activated );
[…]
