### 2017-11-30 eba8c5c
XML serialization and deserialization added for IFC4Add2. It is possible to specify which configuration should be used for serialization but xBIM will use IFC4Add2 configuration as the default one from now on.
