+++ Xbim.IO/Xml/XbimXmlReader4.cs
@@ -155,6 +155,11 @@ namespace Xbim.IO.Xml
                         header.FileSchema.Schemas.Add("IFC4");
                         break;
                     }
+                    if (input.Value == "http://www.buildingsmart-tech.org/ifcXML/IFC4/Add2")
+                    {
+                        header.FileSchema.Schemas.Add("IFC4");
+                        break;
+                    }
                 }
                 input.MoveToElement();
             }            
@@ -680,6 +685,9 @@ namespace Xbim.IO.Xml
 
                     if (string.Equals(input.Value, "http://www.buildingsmart-tech.org/ifcXML/IFC4/final", StringComparison.OrdinalIgnoreCase))
                         return XmlSchemaVersion.Ifc4;
+
+                    if (string.Equals(input.Value, "http://www.buildingsmart-tech.org/ifcXML/IFC4/Add2", StringComparison.OrdinalIgnoreCase))
+                        return XmlSchemaVersion.Ifc4Add2;
                 }
                 input.MoveToElement();
             }
@@ -692,6 +700,7 @@ namespace Xbim.IO.Xml
         // ReSharper disable once InconsistentNaming
         Ifc2x3,
         Ifc4Add1,
+        Ifc4Add2,
         Ifc4,
         Unknown
     }
