+++ Xbim.Ifc2x3/Interfaces/IFC4/IfcPoint.cs
@@ -17,8 +17,8 @@ namespace Xbim.Ifc2x3.GeometryResource
 {
 	public partial class @IfcPoint : IIfcPoint
 	{
-        //## Custom code
+	//## Custom code
         Ifc4.GeometryResource.IfcDimensionCount Ifc4.GeometricModelResource.IfcGeometricSetSelect.Dim { get { return new Ifc4.GeometryResource.IfcDimensionCount(Dim); } }
         //##
-    }
+	}
 }
