### 2017-11-30 5e25a75
Minor docs update.
