+++ README.md
@@ -1,46 +1,62 @@
 # XbimEssentials
-Part of Xbim; the eXtensible [Building Information Modelling (BIM) Toolkit](https://xbimteam.github.io/)
+XbimEssentials is part of Xbim, the eXtensible [Building Information Modelling Toolkit](https://xbimteam.github.io/)
 
-Build Status (master branch): [ ![Build Status](http://xbimbuilds.cloudapp.net/app/rest/builds/buildType:(id:XbimEssentials_XbimEssentials),branch:(name:master)/statusIcon "Build Status") ](http://xbimbuilds.cloudapp.net/project.html?projectId=XbimEssentials&tab=projectOverview "Build Status")
+## Build Status
 
-Build Status (develop branch): [ ![Build Status](http://xbimbuilds.cloudapp.net/app/rest/builds/buildType:(id:XbimEssentials_XbimEssentials),branch:(name:develop)/statusIcon "Build Status") ](http://xbimbuilds.cloudapp.net/project.html?projectId=XbimEssentials&tab=projectOverview "Build Status")
+Master: [ ![Build Status](http://xbimbuilds.cloudapp.net/app/rest/builds/buildType:(id:XbimEssentials_XbimEssentials),branch:(name:master)/statusIcon "Build Status") ](http://xbimbuilds.cloudapp.net/project.html?projectId=XbimEssentials&tab=projectOverview "Build Status")
+
+Develop branch: [ ![Build Status](http://xbimbuilds.cloudapp.net/app/rest/builds/buildType:(id:XbimEssentials_XbimEssentials),branch:(name:develop)/statusIcon "Build Status") ](http://xbimbuilds.cloudapp.net/project.html?projectId=XbimEssentials&tab=projectOverview "Build Status")
+
+The XbimTeam is committed to posting code that always compiles, if you see a build failure here the error is generally due to omitted increases in the Nuget version number.
 
 ## What is it?
 
-The xBIM Tookit (eXtensible Building Information Modelling) is an open-source, software development BIM toolkit that 
+The xBIM Tookit (eXtensible Building Information Modelling) is an open-source, software development BIM toolkit that
 supports the BuildingSmart Data Model (aka the [Industry Foundation Classes IFC](http://en.wikipedia.org/wiki/Industry_Foundation_Classes)).
 
-xBIM allows developers to read, create and view [Building Information (BIM)](http://en.wikipedia.org/wiki/Building_information_modeling) Models in the IFC format. 
-There is full support for geometric, topological operations and visualisation. In addition xBIM supports 
+xBIM allows developers to read, create and view [Building Information (BIM)](http://en.wikipedia.org/wiki/Building_information_modeling) Models in the IFC format.
+There is full support for geometric, topological operations and visualisation. In addition xBIM supports
 bi-directional translation between IFC and COBie formats
 
 ## Getting Started
 
+### Using the library
+
+If you with to use Xbim.Essential in your code you just have to add the package to your Solution in Visual Studio using Nuget's Package Manager Console and issuing the following command:
+
+```
+PM> Install-Package Xbim.Essentials
+```
+
+### Compilation
+
 You will need Visual Studio 2013 or newer to compile the Solution. All solutions target .NET 4.0. The 4.0 Client profile
 may be supported for some projects. The roadmap expects to move to 4.5 versions of the .NET framework soon.
 
-XbimEssentials is a software library to be used for the creation of complex applications, other repositories under the [XbimTeam](https://github.com/xBimTeam) page include a number of example applications to demonstrate its capabilities
+### Examples
 
-* [XbimXplorer](https://github.com/xBimTeam/XbimWindowsUI) - a Windows WPF sample application that can open and render 3D IFC models (and native XBIM models ) as well as displaying semantic data.
[…]
