+++ CONTRIBUTING.md
@@ -2,21 +2,21 @@
 
 ## **Did you find a bug?**
 
-* Ensure the bug was not already reported by searching on GitHub under [Issues](https://github.com/xBimTeam/XbimEssentials/issues). 
+* Ensure the bug was not already reported by searching on GitHub under [Issues](https://github.com/xBimTeam/XbimEssentials/issues).
 
-* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/xBimTeam/XbimEssentials/issues/new). 
-Be sure to include a **title and clear description**, as much relevant information as possible; if useful an IFC file should be attached to 
+* If you're unable to find an open issue addressing the problem, [open a new one](https://github.com/xBimTeam/XbimEssentials/issues/new).
+Be sure to include a **title and clear description**, as much relevant information as possible; if useful an IFC file should be attached to
 the issue and a **code sample** demonstrating the fault. xBIM can be used to extract parts of IFC file to isolate the problem if it is data-related.
 
 * The Xbim toolkit is composed of a collection of projects, please ensure that you post your issue under the relevant project.
 
 ## **Did you write a patch that fixes a bug?**
 
-* Open a new [GitHub pull](https://help.github.com/articles/about-pull-requests/) request with the patch.
+* Open a new [GitHub pull request](https://help.github.com/articles/about-pull-requests/) with the patch.
 
-* we have two branches in the project, please work on the development branch; your contribution will be merged there and then to master upon the next release.
+* we have two branches in the project, **please work on the development branch**; your contribution will be merged there and then to master upon the next release.
 
-* Ensure the PR description clearly describes the problem and solution. Include a relevant issue number if applicable.
+* Ensure the PR description clearly captures the problem and solution. Include a relevant issue number, if applicable.
 
 ## **Did you fix whitespace, format code, or make a purely cosmetic patch?**
 
@@ -32,6 +32,6 @@ We are still likely to accept your PR in this case, patches that improve the doc
 
 Xbim is a volunteer effort. We encourage you to pitch in!
 
-Thanks! 
+Thanks!
 
 The Xbim Team
