### 2017-11-30 7200a59
Joss paper update: fixed references and improved image for DLL dependency diagram.
Restored original title, minor changes in punctuation.
