+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Xbim.Essentials: full control over Building Information Models'
+title: 'Xbim.Essentials: a library for interoperable building information applications'
 tags:
   - BIM
   - BuildingSmart
@@ -17,7 +17,7 @@ authors:
 affiliations:
  - name: Northumbria University
    index: 1
-date: 04 September 2017
+date: 4^th^ September 2017
 bibliography: paper.bib
 ...
 
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 The Xbim.Essentials library offers data extraction, data transformation and data validation functions for Building Information Models (BIM); its robust and optimised implementation allows efficient operations on gigabyte-range files for researchers and practitioners interested in the built environment.
 
-The library implements the complete object model of BuildingSMART's IFC schemas [-@buildingsmart_international_summary_2017] along with APIs for their management under the terms of the CDDL [@sun_microsystems_common_2005], which makes it also suitable for commercial research and development projects.
+The library implements the complete object model of BuildingSMART's IFC schemas [-@buildingSMARTSummaryIFCReleases2017] along with APIs for their management under the terms of the CDDL [@SunMicrosystemsCommonDevelopmentDistribution2005], which makes it also suitable for commercial research and development projects.
 
 Relevant fields of research span from social sciences to construction engineering and applied mathematics on account of the breadth of domains covered in the over 800 classes of the schema across building lifecycles.
 
@@ -39,26 +39,22 @@ The project implements public APIs for:
 Architectural features of the solution include:
 
 - Full access to the models through C# interfaces, allowing schema-agnostic data logic on any supported version.
-- Disk-based and in-memory options to suit diverse of workflow infrastructures.
+- Disk-based and in-memory options to suit diverse workflow infrastructures.
 - Log management for notification events where exceptions are not appropriate.
 
-Additionally, the Xbim.IO.TableStore namespace supports similar APIs, when conceptually suitable, on BuildingSMART's COBie object model [@buildingsmart_cobie_2013], extending the domain of relevance to construction operations and facility management.
+Additionally, the Xbim.IO.TableStore namespace supports similar APIs, when conceptually suitable, on BuildingSMART's COBie object model [@buildingSMARTCOBieModelview2013], extending the domain of relevance to construction operations and facility management.
 
-![Figure 1: Internal dependencies graph.](CodeDependenciesGraph.png)
+![Internal DLL dependencies graph.](RelDiagram.png)
 
-Where appropriate, classes in the codebase are generated programmatically from the formal EXPRESS specification files [@iso_iso-10303-11:2004_2013] ensuring full compliance with the standards.
+Where appropriate, classes in the codebase are generated programmatically from the formal EXPRESS specification files [@ISOISO10303112004Industrial2013] ensuring full compliance with the standards.
 
-Schemas currently implemented are:
+Schemas currently implemented are: 'IFC2x3 Technical Corrigendum 1' [@ModelSupportGroupIFC2x3TechnicalCorrigendum2007], 'IFC4 Addendum 1' [@ModelSupportGroupIFC4AddendumFinal2015] and 'COBie MVD' [@buildingSMARTCOBieModelview2013].
 
-- IFC2x3 Technical Corrigendum 1 [@model_support_group_ifc2x3_2007]
-- IFC4 Addendum 1 [@model_support_group_ifc4_2015]
-- COBie MVD [@buildingsmart_cobie_2013]
[…]
