+++ paper/paper.bib
@@ -1,170 +1,151 @@
-@online{buildingsmart_international_summary_2017,
+@legislation{ModelSupportGroup,
+  timestamp = {2017-06-01T08:31:23Z},
+  title = {Model {{Support Group}} of {{buildingSMART International}}},
+  url = {http://www.buildingsmart-tech.org/about-us/msg},
+  groups = {MVD and IDM}
+}
+
+@online{buildingSMARTCOBieModelview2013,
+  timestamp = {2017-09-04T17:54:02Z},
+  title = {{{COBie}} - {{Model}} View Definition},
+  url = {http://docs.buildingsmartalliance.org/MVD_COBIE/},
+  author = {{buildingSMART}},
+  urldate = {2017-08-07},
+  date = {2013-10-30},
+  groups = {COBie,Standards}
+}
+
+@misc{ISOISO10303112004Industrial2013,
+  timestamp = {2017-09-04T16:23:07Z},
+  title = {{{ISO}}-10303-11:2004 {{Industrial}} Automation Systems and Integration -- {{Product}} Data Representation and Exchange -- {{Part}} 11: {{Description}} Methods: {{The EXPRESS}} Language Reference Manual},
+  url = {https://www.iso.org/standard/38047.html},
+  shorttitle = {{{EXPRESS}} Data Modeling Language},
+  abstract = {ISO 10303 specifies a language by which aspects of product data can be defined. The language is called EXPRESS. ISO 10303-11:2004 also specifies a graphical representation for a subset of the constructs in the EXPRESS language. This graphical representation is called EXPRESS-G. EXPRESS is a data specification language as defined in ISO 10303-1. It consists of language elements that allow an unambiguous data definition and specification of constraints on the data defined.
+
+The following are within the scope of ISO 10303-11:2004:
+
+data types;
+constraints on instances of the data types.
+The following are outside the scope of this part of ISO 10303:
+
+definition of database formats;
+definition of file formats;
+definition of transfer formats;
+process control;
+information processing;
+exception handling.
+EXPRESS is not a programming language.},
+  publisher = {{International Organization for Standardization}},
+  author = {{ISO}},
+  urldate = {2017-09-04},
+  date = {2013-04},
+  groups = {Standards}
+}
+
+@online{buildingSMARTSummaryIFCReleases2017,
+  timestamp = {2017-11-26T11:20:21Z},
   title = {Summary of {{IFC Releases}}},
   url = {http://www.buildingsmart-tech.org/specifications/ifc-releases/summary},
   abstract = {Overview about IFC releases, including current IFC release, background information about previous IFC releases, and preview of future IFC releases.},
-  timestamp = {2017-09-04T23:38:45Z},
   journaltitle = {buildingSMART-Tech.org},
   type = {Technical Documentation},
-  author = {{buildingSMART International}},
+  author = {{buildingSMART}},
   urldate = {2017-09-04},
   date = {2017},
-  groups = {Institutions and standards,Institutions and standards}
+  groups = {Standards}
 }
 
-@misc{sun_microsystems_common_2005,
+@online{ModelSupportGroupIFC2x3TechnicalCorrigendum2007,
+  timestamp = {2017-09-04T17:53:08Z},
+  title = {{{IFC2x3 Technical Corrigendum}} 1 - {{Final Documentation}}},
+  url = {http://www.buildingsmart-tech.org/ifc/IFC2x3/TC1/html/index.htm},
[…]
