### 2017-11-30 #473 @CBenghi
@Haacked,
I've noticed that you have skipped some checkboxes above, so I've improved the submission as follows:

- **Example usage**: I've added an example section in the readme.md to point to the examples available on docs.xbim.net.
- **References**: I've updated the paper.md file to fix two references that were misinterpreted in the previous version.  I've double checked that any entry that does not list a DOI does not have one at all.

Now that I've seen how the paper gets rendered I've also corrected some markdown that worked in my pandoc configuration but has a different behaviour here. While at it I've also changed some text to improve the pagination (a new page happened just before a list) and the diagram of DLL dependencies. I'm not sure how to trigger a recompilation of the PDF.

Have you got any specific guidance for  **Functionality documentation**?

Claudio
