+++ paper/paper.bib
@@ -1,22 +1,17 @@
 @legislation{ModelSupportGroup,
-  timestamp = {2017-06-01T08:31:23Z},
   title = {Model {{Support Group}} of {{buildingSMART International}}},
-  url = {http://www.buildingsmart-tech.org/about-us/msg},
-  groups = {MVD and IDM}
+  url = {http://www.buildingsmart-tech.org/about-us/msg}
 }
 
 @online{buildingSMARTCOBieModelview2013,
-  timestamp = {2017-09-04T17:54:02Z},
   title = {{{COBie}} - {{Model}} View Definition},
   url = {http://docs.buildingsmartalliance.org/MVD_COBIE/},
   author = {{buildingSMART}},
   urldate = {2017-08-07},
-  date = {2013-10-30},
-  groups = {COBie,Standards}
+  date = {2013-10-30}
 }
 
 @misc{ISOISO10303112004Industrial2013,
-  timestamp = {2017-09-04T16:23:07Z},
   title = {{{ISO}}-10303-11:2004 {{Industrial}} Automation Systems and Integration -- {{Product}} Data Representation and Exchange -- {{Part}} 11: {{Description}} Methods: {{The EXPRESS}} Language Reference Manual},
   url = {https://www.iso.org/standard/38047.html},
   shorttitle = {{{EXPRESS}} Data Modeling Language},
@@ -38,114 +33,94 @@ EXPRESS is not a programming language.},
   publisher = {{International Organization for Standardization}},
   author = {{ISO}},
   urldate = {2017-09-04},
-  date = {2013-04},
-  groups = {Standards}
+  date = {2013-04}
 }
 
 @online{buildingSMARTSummaryIFCReleases2017,
-  timestamp = {2017-11-26T11:20:21Z},
   title = {Summary of {{IFC Releases}}},
   url = {http://www.buildingsmart-tech.org/specifications/ifc-releases/summary},
   abstract = {Overview about IFC releases, including current IFC release, background information about previous IFC releases, and preview of future IFC releases.},
-  journaltitle = {buildingSMART-Tech.org},
   type = {Technical Documentation},
+  journaltitle = {buildingSMART-Tech.org},
   author = {{buildingSMART}},
   urldate = {2017-09-04},
-  date = {2017},
-  groups = {Standards}
+  date = {2017}
 }
 
 @online{ModelSupportGroupIFC2x3TechnicalCorrigendum2007,
-  timestamp = {2017-09-04T17:53:08Z},
   title = {{{IFC2x3 Technical Corrigendum}} 1 - {{Final Documentation}}},
   url = {http://www.buildingsmart-tech.org/ifc/IFC2x3/TC1/html/index.htm},
   author = {{Model Support Group}},
   urldate = {2017-09-04},
-  date = {2007},
-  groups = {Standards}
+  date = {2007}
 }
 
 @online{ModelSupportGroupIFC4AddendumFinal2015,
-  timestamp = {2017-09-04T16:39:48Z},
   title = {{{IFC4 Addendum}} 1 - {{Final}} Documentation},
   url = {http://www.buildingsmart-tech.org/ifc/IFC4/Add1/html/},
   author = {{Model Support Group}},
   urldate = {2017-09-04},
-  date = {2015},
-  groups = {Standards}
+  date = {2015}
 }
 
 @online{InnovateUKTier2Tiercollaborationinterface2015,
-  timestamp = {2017-09-04T22:33:20Z},
   title = {{{Tier2Tier A}} Collaboration Interface between Construction Main Contractors and Their Supply Chain Specialist Sub-Contractors},
[…]
