### 2017-12-02 #473 @Haacked
> I've noticed that you have skipped some checkboxes above

@CBenghi, sorry about that. I meant to get to it but I'm at a conference in the Dominican Republic so I've been very busy and didn't get a chance to finish.

I wasn't sure how to check that the DOI references are correct.

> Have you got any specific guidance for Functionality documentation?

I mostly looked at the `///` style code comments for public APIs in the code. I think it could use more of them, but it looks like the important stuff is document. Is there API documentation I'm missing?
