### 2017-12-02 #473 @CBenghi
@Haacked,

my apologies, I had no intention to put pressure on you. There's absolutely no rush, I was only trying to improve the submission so that it would be easier for you to review.

We have no external API docs yet. We have documented the APIs with IntelliSense as you have seen, particularly in the areas where most questions and issues have been received. In our experience this seems Ok for people to start engaging, but I agree that there could be more. I'll check again to see if there are obvious omissions tomorrow.

We have plans to automate the IntelliSense comments on auto-generated classes from the ISO documentation (about 1500 classes and 8000 properties) but it might end up in the next release, and - in a way - it could be considered outside the scope of our contribution, these are only implementations of external schemas.

Regarding the DOI:
It's very easy to test that the DOI that we have specified is correct: click on the DOI part of the reference and it'll take you to the a page where you can control the match.

Conversely, ensuring that a registered DOIs do not exist for the others is much harder and I cannot be certain that I've made no mistakes there, but I've [spent hours searching](https://www.crossref.org/guestquery/) with no luck.

I hope you are enjoying the conference.

Best,
Claudio
