### 2017-12-05 #473 @arfon
> Ok, I checked everything off. Is there anything left for me to do?

Great. Thanks @Haacked - just your confirmation that you're happy with the software you've reviewed here.

Once I have that I can proceed with accepting the submission.
