### 2017-12-07 #473 @CBenghi
@arfon,
I've issued a release and recorded it with zenodo with DOI: 10.5281/zenodo.1095192

@Haacked,
very many thanks for your support with the review process. Hugely appreciated.

The very process of submission have made us think hard about the quality of the software including the aspects of community engagement and documentation. Hopefully this practice will spread and we'll get more from the open source movement than ever before.
I'd be happy to help as a reviewer in the future, if useful.

Thanks, and keep up the good work!
